Revert "Refactor RQL configuration and result handling"
This reverts commit b11b8e8a108e3b1811f1d1e8752ee335d2f10376.
