--- src/SoftwareOne.Rql.Abstractions/Operators/IActualOperator.cs
@@ -1,4 +1,0 @@
-﻿namespace SoftwareOne.Rql.Abstractions.Operators
-{
-    public interface IActualOperator { }
-}
