--- src/SoftwareOne.Rql.Abstractions/Operators/IOperator.cs
@@ -1,5 +1,0 @@
-﻿namespace SoftwareOne.Rql.Abstractions.Operators;
-
-public interface IOperator
-{
-}
