+++ src/SoftwareOne.Rql.Linq/Core/IPathInfoBuilder.cs
@@ -1,14 +1,15 @@
 ﻿using SoftwareOne.Rql.Abstractions;
-using SoftwareOne.Rql.Abstractions.Result;
+using SoftwareOne.Rql.Linq.Core.Result;
 using System.Linq.Expressions;
 
-namespace SoftwareOne.Rql.Linq.Core;
-
-internal interface IPathInfoBuilder
+namespace SoftwareOne.Rql.Linq.Core
 {
-    Result<MemberPathInfo> Build(Expression root, RqlExpression rqlExpression);
+    internal interface IPathInfoBuilder
+    {
+        Result<MemberPathInfo> Build(Expression root, RqlExpression rqlExpression);
 
-    Result<MemberPathInfo> Build(Expression root, string path);
+        Result<MemberPathInfo> Build(Expression root, string path);
+    }
+
+    internal record MemberPathInfo(string FullPath, ReadOnlyMemory<char> Path, RqlPropertyInfo PropertyInfo, Expression Expression);
 }
-
-internal record MemberPathInfo(string FullPath, ReadOnlyMemory<char> Path, RqlPropertyInfo PropertyInfo, Expression Expression);
