+++ src/SoftwareOne.Rql.Linq/Core/PathInfoBuilder.cs
@@ -3,78 +3,80 @@ using SoftwareOne.Rql.Abstractions.Argument;
 using SoftwareOne.Rql.Abstractions.Argument.Pointer;
 using SoftwareOne.Rql.Abstractions.Result;
 using SoftwareOne.Rql.Linq.Core.Metadata;
+using SoftwareOne.Rql.Linq.Core.Result;
 using SoftwareOne.Rql.Linq.Services.Context;
 using System.Linq.Expressions;
 
-namespace SoftwareOne.Rql.Linq.Core;
-
-internal abstract class PathInfoBuilder : IPathInfoBuilder
+namespace SoftwareOne.Rql.Linq.Core
 {
-    private readonly IMetadataProvider _metadataProvider;
-    private readonly IBuilderContext _builderContext;
-
-    protected PathInfoBuilder(IMetadataProvider metadataProvider, IBuilderContext builderContext)
+    internal abstract class PathInfoBuilder : IPathInfoBuilder
     {
-        _metadataProvider = metadataProvider;
-        _builderContext = builderContext;
-    }
+        private readonly IMetadataProvider _metadataProvider;
+        private readonly IBuilderContext _builderContext;
 
-    public Result<MemberPathInfo> Build(Expression root, RqlExpression rqlExpression)
-    {
-        switch (rqlExpression)
+        protected PathInfoBuilder(IMetadataProvider metadataProvider, IBuilderContext builderContext)
         {
-            case RqlSelf self:
-                {
-                    if (self.Inner != null)
-                        return Build(root, self.Inner);
-
-                    var path = string.Empty;
-                    return new MemberPathInfo(path, path.AsMemory(0, 0), RqlPropertyInfo.Root, root);
-                }
-            case RqlConstant constant:
-                {
-                    var path = Build(root, constant.Value);
-                    if (path.IsError)
-                        return path.Errors;
-                    return path.Value!;
-                }
-            default:
-                return Error.Validation("Unsupported property node.");
+            _metadataProvider = metadataProvider;
+            _builderContext = builderContext;
         }
-    }
 
-    public Result<MemberPathInfo> Build(Expression root, string path)
-    {
-        var nameSegments = path.Split('.');
-        var aggregatedInfo = nameSegments.Aggregate(
-            new Result<MemberPathInfo>(new MemberPathInfo(path, path.AsMemory(0, 0), null!, root)),
-            (current, segment) =>
+        public Result<MemberPathInfo> Build(Expression root, RqlExpression rqlExpression)
+        {
+            switch (rqlExpression)
             {
-                if (current.IsError)
-                    return current;
+                case RqlSelf self:
+                    {
+                        if (self.Inner != null)
+                            return Build(root, self.Inner);
 
-                var previousLength = current.Value!.Path.Length;
-                var cumulativePath = current.Value.FullPath.AsMemory(0, (previousLength > 0 ? previousLength + 1 : previousLength) + segment.Length);
+                        var path = string.Empty;
+                        return new MemberPathInfo(path, path.AsMemory(0, 0), RqlPropertyInfo.Root, root);
+                    }
+                case RqlConstant constant:
+                    {
+                        var path = Build(root, constant.Value);
+                        if (path.IsError)
+                            return path.Errors;
+                        return path.Value!;
+                    }
+                default:
+                    return Error.Validation("Unsupported property node.");
+            }
+        }
 
-                if (!_metadataProvider.TryGetPropertyByDisplayName(current.Value.Expression.Type, segment, out var propInfo) || propInfo!.IsIgnored)
-                    return Error.Validation("Invalid property path.", _builderContext.GetFullPath(cumulativePath.ToString()));
+        public Result<MemberPathInfo> Build(Expression root, string path)
+        {
+            var nameSegments = path.Split('.');
+            var aggregatedInfo = nameSegments.Aggregate(
+                new Result<MemberPathInfo>(new MemberPathInfo(path, path.AsMemory(0, 0), null!, root)),
+                (current, segment) =>
+                {
+                    if (current.IsError)
+                        return current;
 
-                var expression = (Expression)Expression.MakeMemberAccess(current.Value!.Expression, propInfo!.Property!);
-                var pathInfo = new MemberPathInfo(current.Value.FullPath, cumulativePath, propInfo, expression);
+                    var previousLength = current.Value!.Path.Length;
+                    var cumulativePath = current.Value.FullPath.AsMemory(0, (previousLength > 0 ? previousLength + 1 : previousLength) + segment.Length);
 
-                var validationResult = ValidatePath(pathInfo);
+                    if (!_metadataProvider.TryGetPropertyByDisplayName(current.Value.Expression.Type, segment, out var propInfo) || propInfo!.IsIgnored)
+                        return Error.Validation("Invalid property path.", _builderContext.GetFullPath(cumulativePath.ToString()));
 
-                if (validationResult.IsError)
-                    return validationResult.Errors;
+                    var expression = (Expression)Expression.MakeMemberAccess(current.Value!.Expression, propInfo!.Property!);
+                    var pathInfo = new MemberPathInfo(current.Value.FullPath, cumulativePath, propInfo, expression);
 
-                return pathInfo;
-            });
+                    var validationResult = ValidatePath(pathInfo);
 
-        if (aggregatedInfo.IsError)
-            return aggregatedInfo.Errors;
+                    if (validationResult.IsError)
+                        return validationResult.Errors;
 
-        return aggregatedInfo.Value!;
+                    return pathInfo;
+                });
+
+            if (aggregatedInfo.IsError)
+                return aggregatedInfo.Errors;
+
+            return aggregatedInfo.Value!;
+        }
+
+        protected abstract Result<bool> ValidatePath(MemberPathInfo pathInfo);
     }
-
-    protected abstract Result<bool> ValidatePath(MemberPathInfo pathInfo);
 }
