+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Builders/IConcreteExpressionBuilder.cs
@@ -1,5 +1,5 @@
 ﻿using SoftwareOne.Rql.Abstractions;
-using SoftwareOne.Rql.Abstractions.Result;
+using SoftwareOne.Rql.Linq.Core.Result;
 using System.Linq.Expressions;
 
 namespace SoftwareOne.Rql.Linq.Services.Filtering.Builders;
