+++ src/SoftwareOne.Rql.Linq/Services/Filtering/FilteringPathInfoBuilder.cs
@@ -1,27 +1,29 @@
 ﻿using SoftwareOne.Rql.Abstractions.Result;
 using SoftwareOne.Rql.Linq.Core;
 using SoftwareOne.Rql.Linq.Core.Metadata;
+using SoftwareOne.Rql.Linq.Core.Result;
 using SoftwareOne.Rql.Linq.Services.Context;
 
-namespace SoftwareOne.Rql.Linq.Services.Filtering;
-
-internal interface IFilteringPathInfoBuilder : IPathInfoBuilder { }
-
-internal class FilteringPathInfoBuilder : PathInfoBuilder, IFilteringPathInfoBuilder
+namespace SoftwareOne.Rql.Linq.Services.Filtering
 {
-    private readonly IActionValidator _actionValidator;
-    private readonly IBuilderContext _builderContext;
+    internal interface IFilteringPathInfoBuilder : IPathInfoBuilder { }
 
-    public FilteringPathInfoBuilder(IActionValidator actionValidator, IMetadataProvider metadataProvider, IBuilderContext builderContext) : base(metadataProvider, builderContext)
+    internal class FilteringPathInfoBuilder : PathInfoBuilder, IFilteringPathInfoBuilder
     {
-        _actionValidator = actionValidator;
-        _builderContext = builderContext;
-    }
+        private readonly IActionValidator _actionValidator;
+        private readonly IBuilderContext _builderContext;
 
-    protected override Result<bool> ValidatePath(MemberPathInfo pathInfo)
-    {
-        if (!_actionValidator.Validate(pathInfo.PropertyInfo, RqlActions.Filter))
-            return Error.Validation("Filtering is not permitted.", _builderContext.GetFullPath(pathInfo.Path.ToString()));
-        return true;
+        public FilteringPathInfoBuilder(IActionValidator actionValidator, IMetadataProvider metadataProvider, IBuilderContext builderContext) : base(metadataProvider, builderContext)
+        {
+            _actionValidator = actionValidator;
+            _builderContext = builderContext;
+        }
+
+        protected override Result<bool> ValidatePath(MemberPathInfo pathInfo)
+        {
+            if (!_actionValidator.Validate(pathInfo.PropertyInfo, RqlActions.Filter))
+                return Error.Validation("Filtering is not permitted.", _builderContext.GetFullPath(pathInfo.Path.ToString()));
+            return true;
+        }
     }
 }
