+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Operators/Collection/IAll.cs
@@ -1,5 +1,4 @@
 ﻿using SoftwareOne.Rql.Abstractions.Collection;
-using SoftwareOne.Rql.Abstractions.Operators;
 using SoftwareOne.Rql.Linq.Services.Filtering.Operators.Collection.Implementation;
 
 namespace SoftwareOne.Rql.Linq.Services.Filtering.Operators.Collection;
