+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Operators/Collection/ICollectionOperator.cs
@@ -1,6 +1,5 @@
 ﻿using SoftwareOne.Rql.Abstractions;
-using SoftwareOne.Rql.Abstractions.Operators;
-using SoftwareOne.Rql.Abstractions.Result;
+using SoftwareOne.Rql.Linq.Core.Result;
 using System.Linq.Expressions;
 
 namespace SoftwareOne.Rql.Linq.Services.Filtering.Operators.Collection;
