+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Operators/Collection/Implementation/Any.cs
@@ -1,4 +1,4 @@
-﻿using SoftwareOne.Rql.Abstractions.Result;
+﻿using SoftwareOne.Rql.Linq.Core.Result;
 using System.Reflection;
 
 namespace SoftwareOne.Rql.Linq.Services.Filtering.Operators.Collection.Implementation;
