+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Operators/Comparison/IComparisonOperator.cs
@@ -1,7 +1,8 @@
-﻿using SoftwareOne.Rql.Abstractions.Result;
+﻿using SoftwareOne.Rql.Abstractions;
+using SoftwareOne.Rql.Linq.Core.Result;
 using System.Linq.Expressions;
 
-namespace SoftwareOne.Rql.Abstractions.Operators;
+namespace SoftwareOne.Rql.Linq.Services.Filtering.Operators.Comparison;
 
 public interface IComparisonOperator : IOperator
 {
