+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Operators/Comparison/ILessThanOrEqual.cs
@@ -1,5 +1,4 @@
 ﻿using SoftwareOne.Rql.Abstractions.Binary;
-using SoftwareOne.Rql.Abstractions.Operators;
 using SoftwareOne.Rql.Linq.Services.Filtering.Operators.Comparison.Implementation;
 
 namespace SoftwareOne.Rql.Linq.Services.Filtering.Operators.Comparison;
