+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Operators/Comparison/Implementation/ComparisonOperator.cs
@@ -1,9 +1,9 @@
 ﻿using SoftwareOne.Rql.Abstractions;
 using SoftwareOne.Rql.Abstractions.Configuration;
 using SoftwareOne.Rql.Abstractions.Configuration.Filter;
-using SoftwareOne.Rql.Abstractions.Operators;
 using SoftwareOne.Rql.Abstractions.Result;
 using SoftwareOne.Rql.Linq.Core.Expressions;
+using SoftwareOne.Rql.Linq.Core.Result;
 using System.Linq.Expressions;
 using System.Reflection;
