+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Operators/IActualOperator.cs
@@ -1,0 +1,4 @@
+﻿namespace SoftwareOne.Rql.Linq.Services.Filtering.Operators
+{
+    public interface IActualOperator { }
+}
