+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Operators/IOperator.cs
@@ -1,0 +1,5 @@
+﻿namespace SoftwareOne.Rql.Linq.Services.Filtering.Operators;
+
+public interface IOperator
+{
+}
