+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Operators/IOperatorHandlerProvider.cs
@@ -1,6 +1,4 @@
-﻿using SoftwareOne.Rql.Abstractions.Operators;
-
-namespace SoftwareOne.Rql.Linq.Services.Filtering.Operators;
+﻿namespace SoftwareOne.Rql.Linq.Services.Filtering.Operators;
 
 public interface IOperatorHandlerProvider
 {
