+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Operators/List/IListIn.cs
@@ -1,5 +1,4 @@
 ﻿using SoftwareOne.Rql.Abstractions.Binary;
-using SoftwareOne.Rql.Abstractions.Operators;
 using SoftwareOne.Rql.Linq.Services.Filtering.Operators.List.Implementation;
 
 namespace SoftwareOne.Rql.Linq.Services.Filtering.Operators.List;
