+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Operators/List/Implementation/ListIn.cs
@@ -1,6 +1,7 @@
 ﻿using SoftwareOne.Rql.Abstractions;
 using SoftwareOne.Rql.Abstractions.Result;
 using SoftwareOne.Rql.Linq.Core.Expressions;
+using SoftwareOne.Rql.Linq.Core.Result;
 using System.Collections;
 using System.Linq.Expressions;
 using System.Reflection;
