+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Operators/Search/ISearchOperator.cs
@@ -1,8 +1,8 @@
 ﻿using SoftwareOne.Rql.Abstractions;
-using SoftwareOne.Rql.Abstractions.Result;
+using SoftwareOne.Rql.Linq.Core.Result;
 using System.Linq.Expressions;
 
-namespace SoftwareOne.Rql.Abstractions.Operators;
+namespace SoftwareOne.Rql.Linq.Services.Filtering.Operators.Search;
 
 public interface ISearchOperator : IOperator
 {
