+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Operators/Unary/IUnaryOperator.cs
@@ -1,5 +1,4 @@
-﻿using SoftwareOne.Rql.Abstractions.Operators;
-using SoftwareOne.Rql.Abstractions.Result;
+﻿using SoftwareOne.Rql.Linq.Core.Result;
 using System.Linq.Expressions;
 
 namespace SoftwareOne.Rql.Linq.Services.Filtering.Operators.Unary;
