+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Operators/Unary/Implementation/Not.cs
@@ -1,4 +1,4 @@
-﻿using SoftwareOne.Rql.Abstractions.Result;
+﻿using SoftwareOne.Rql.Linq.Core.Result;
 using System.Linq.Expressions;
 
 namespace SoftwareOne.Rql.Linq.Services.Filtering.Operators.Unary.Implementation;
