+++ src/SoftwareOne.Rql.Linq/Services/Filtering/Operators/ValidationHelper.cs
@@ -1,5 +1,6 @@
 ﻿using SoftwareOne.Rql.Abstractions;
 using SoftwareOne.Rql.Abstractions.Result;
+using SoftwareOne.Rql.Linq.Core.Result;
 
 namespace SoftwareOne.Rql.Linq.Services.Filtering.Operators;
 internal static class ValidationHelper
