+++ src/SoftwareOne.Rql.Linq/Services/Ordering/OrderingPathInfoBuilder.cs
@@ -1,27 +1,29 @@
 ﻿using SoftwareOne.Rql.Abstractions.Result;
 using SoftwareOne.Rql.Linq.Core;
 using SoftwareOne.Rql.Linq.Core.Metadata;
+using SoftwareOne.Rql.Linq.Core.Result;
 using SoftwareOne.Rql.Linq.Services.Context;
 
-namespace SoftwareOne.Rql.Linq.Services.Ordering;
-
-internal interface IOrderingPathInfoBuilder : IPathInfoBuilder { }
-
-internal class OrderingPathInfoBuilder : PathInfoBuilder, IOrderingPathInfoBuilder
+namespace SoftwareOne.Rql.Linq.Services.Ordering
 {
-    private readonly IActionValidator _actionValidator;
-    private readonly IBuilderContext _builderContext;
+    internal interface IOrderingPathInfoBuilder : IPathInfoBuilder { }
 
-    public OrderingPathInfoBuilder(IActionValidator actionValidator, IMetadataProvider metadataProvider, IBuilderContext builderContext) : base(metadataProvider, builderContext)
+    internal class OrderingPathInfoBuilder : PathInfoBuilder, IOrderingPathInfoBuilder
     {
-        _actionValidator = actionValidator;
-        _builderContext = builderContext;
-    }
+        private readonly IActionValidator _actionValidator;
+        private readonly IBuilderContext _builderContext;
 
-    protected override Result<bool> ValidatePath(MemberPathInfo pathInfo)
-    {
-        if (!_actionValidator.Validate(pathInfo.PropertyInfo, RqlActions.Order))
-            return Error.Validation("Ordering is not permitted.", _builderContext.GetFullPath(pathInfo.Path.ToString()));
-        return true;
+        public OrderingPathInfoBuilder(IActionValidator actionValidator, IMetadataProvider metadataProvider, IBuilderContext builderContext) : base(metadataProvider, builderContext)
+        {
+            _actionValidator = actionValidator;
+            _builderContext = builderContext;
+        }
+
+        protected override Result<bool> ValidatePath(MemberPathInfo pathInfo)
+        {
+            if (!_actionValidator.Validate(pathInfo.PropertyInfo, RqlActions.Order))
+                return Error.Validation("Ordering is not permitted.", _builderContext.GetFullPath(pathInfo.Path.ToString()));
+            return true;
+        }
     }
 }
