+++ tests/Rql.Tests.Integration/Tests/Extensibility/Utility/CustomEqualHandler.cs
@@ -1,14 +1,15 @@
 ﻿using SoftwareOne.Rql.Abstractions;
-using SoftwareOne.Rql.Abstractions.Result;
+using SoftwareOne.Rql.Linq.Core.Result;
 using SoftwareOne.Rql.Linq.Services.Filtering.Operators.Comparison;
 using System.Linq.Expressions;
 
-namespace Rql.Tests.Integration.Tests.Extensibility.Utility;
-
-internal class CustomEqualHandler : IEqual
+namespace Rql.Tests.Integration.Tests.Extensibility.Utility
 {
-    public Result<Expression> MakeExpression(IRqlPropertyInfo propertyInfo, Expression member, string? value)
+    internal class CustomEqualHandler : IEqual
     {
-        throw new NotImplementedException();
+        public Result<Expression> MakeExpression(IRqlPropertyInfo propertyInfo, Expression member, string? value)
+        {
+            throw new NotImplementedException();
+        }
     }
 }
