+++ tests/SoftwareOne.Rql.Linq.UnitTests/Filtering/Builders/BinaryExpressionBuilderTests.cs
@@ -2,12 +2,14 @@
 using SoftwareOne.Rql.Abstractions.Argument;
 using SoftwareOne.Rql.Abstractions.Binary;
 using SoftwareOne.Rql.Abstractions.Group;
-using SoftwareOne.Rql.Abstractions.Operators;
 using SoftwareOne.Rql.Abstractions.Result;
 using SoftwareOne.Rql.Linq.Core;
 using SoftwareOne.Rql.Linq.Services.Filtering;
 using SoftwareOne.Rql.Linq.Services.Filtering.Builders;
 using SoftwareOne.Rql.Linq.Services.Filtering.Operators;
+using SoftwareOne.Rql.Linq.Services.Filtering.Operators.Comparison;
+using SoftwareOne.Rql.Linq.Services.Filtering.Operators.List;
+using SoftwareOne.Rql.Linq.Services.Filtering.Operators.Search;
 using System.Linq.Expressions;
 using Xunit;
