+++ tests/SoftwareOne.Rql.Linq.UnitTests/Result/ResultTests.cs
@@ -1,4 +1,5 @@
 ﻿using SoftwareOne.Rql.Abstractions.Result;
+using SoftwareOne.Rql.Linq.Core.Result;
 using Xunit;
 
 namespace SoftwareOne.Rql.Linq.UnitTests.Result;
